requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
aiohttp==3.10.5
aiosqlite==0.20.0
python-dateutil==2.9.0.post0
